requirements.txt
Django==3.1.5
djangorestframework==3.12.4
